refactor: extract secret hygiene inventory seam
