+++ crates/app/src/config/mod.rs
@@ -7,6 +7,7 @@ mod memory;
 mod outbound_http;
 mod provider;
 mod runtime;
+mod secret_hygiene;
 mod shared;
 mod tools;
 
@@ -116,6 +117,10 @@ pub use runtime::{
     preferred_provider_selector, provider_selector_catalog, provider_selector_recommendation_hint,
     render, resolve_provider_selector, supported_validation_locales, validate_file,
     validate_file_with_locale, write, write_template,
+};
+pub use secret_hygiene::{
+    SecretObservation, SecretObservationCounts, SecretReferenceKind, collect_env_pointer_diagnostics,
+    collect_secret_observations, observation_paths_for_kind, summarize_secret_observations,
 };
 pub(crate) use runtime::{normalize_dispatch_account_id, normalize_dispatch_channel_id};
 pub(crate) use shared::ConfigValidationIssue;
