+++ crates/daemon/src/doctor_security_cli.rs
@@ -4,7 +4,6 @@ use std::os::unix::fs::PermissionsExt;
 use std::path::Path;
 
 use loong_app as mvp;
-use loong_contracts::SecretRef;
 use loong_spec::CliResult;
 use serde::Serialize;
 use serde_json::json;
@@ -94,41 +93,6 @@ pub struct DoctorSecurityAuditExecution {
     pub ok: bool,
     pub summary: SecurityAuditSummary,
     pub findings: Vec<SecurityFinding>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SecretReferenceKind {
-    Env,
-    File,
-    Exec,
-    InlineLiteral,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct SecretObservation {
-    field_path: String,
-    kind: SecretReferenceKind,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-struct SecretObservationCounts {
-    env: usize,
-    file: usize,
-    exec: usize,
-    inline_literal: usize,
-}
-
-const PROVIDER_SECRET_HEADER_NAMES: &[&str] = &["authorization", "x-api-key"];
-
-impl SecretObservationCounts {
-    fn record(&mut self, kind: SecretReferenceKind) {
-        match kind {
-            SecretReferenceKind::Env => self.env += 1,
-            SecretReferenceKind::File => self.file += 1,
-            SecretReferenceKind::Exec => self.exec += 1,
-            SecretReferenceKind::InlineLiteral => self.inline_literal += 1,
-        }
-    }
 }
 
 pub async fn run_doctor_security_cli(options: DoctorSecurityCommandOptions) -> CliResult<()> {
@@ -694,14 +658,17 @@ fn assess_secret_hygiene(
     config_path: &Path,
     config: &mvp::config::LoongConfig,
 ) -> CliResult<SecurityFinding> {
-    let mut observations = collect_secret_observations(config);
+    let mut observations = mvp::config::collect_secret_observations(config);
     observations.sort_by(|left, right| left.field_path.cmp(&right.field_path));
 
-    let counts = summarize_secret_observations(&observations);
-    let env_pointer_diagnostics = collect_env_pointer_diagnostics(config);
+    let counts = mvp::config::summarize_secret_observations(&observations);
+    let env_pointer_diagnostics = mvp::config::collect_env_pointer_diagnostics(config);
     let inline_paths =
-        observation_paths_for_kind(&observations, SecretReferenceKind::InlineLiteral);
-    let exec_paths = observation_paths_for_kind(&observations, SecretReferenceKind::Exec);
+        mvp::config::observation_paths_for_kind(&observations, mvp::config::SecretReferenceKind::InlineLiteral);
+    let exec_paths = mvp::config::observation_paths_for_kind(
+        &observations,
+        mvp::config::SecretReferenceKind::Exec,
+    );
 
     let mut evidence = Vec::new();
     let counts_evidence = format!(
@@ -768,7 +735,7 @@ fn assess_secret_hygiene(
         ));
     }
 
-    if counts.exec > 0 || !collect_env_pointer_diagnostics(config).is_empty() {
+    if counts.exec > 0 || !mvp::config::collect_env_pointer_diagnostics(config).is_empty() {
         let summary =
             "Secret references avoid inline literals, but some entries still rely on host exec or env-pointer cleanup."
                 .to_owned();
@@ -779,7 +746,7 @@ fn assess_secret_hygiene(
                     .to_owned(),
             );
         }
-        if !collect_env_pointer_diagnostics(config).is_empty() {
+        if !mvp::config::collect_env_pointer_diagnostics(config).is_empty() {
             next_steps.push(
                 "Normalize env-pointer fields so the config stays on the canonical secret-ref path."
                     .to_owned(),
@@ -887,630 +854,6 @@ fn summarize_findings(findings: &[SecurityFinding]) -> SecurityAuditSummary {
     }
 
     summary
-}
-
-fn collect_env_pointer_diagnostics(
-    config: &mvp::config::LoongConfig,
-) -> Vec<mvp::config::ConfigValidationDiagnostic> {
-    let diagnostics = config.validation_diagnostics();
-    diagnostics
-        .into_iter()
-        .filter(|diagnostic| diagnostic.code.starts_with("config.env_pointer."))
-        .collect()
-}
-
-fn collect_secret_observations(config: &mvp::config::LoongConfig) -> Vec<SecretObservation> {
-    let mut observations = Vec::new();
-
-    collect_provider_secret_observations(config, &mut observations);
-    collect_web_search_secret_observations(config, &mut observations);
-    collect_channel_secret_observations(config, &mut observations);
-
-    observations
-}
-
-fn collect_provider_secret_observations(
-    config: &mvp::config::LoongConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    collect_single_provider_secret_observations("provider", &config.provider, observations);
-
-    for (profile_id, profile) in &config.providers {
-        let field_prefix = format!("providers.{profile_id}");
-        collect_single_provider_secret_observations(
-            field_prefix.as_str(),
-            &profile.provider,
-            observations,
-        );
-    }
-}
-
-fn collect_single_provider_secret_observations(
-    field_prefix: &str,
-    provider: &mvp::config::ProviderConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let api_key_path = format!("{field_prefix}.api_key");
-    push_secret_ref_observation(observations, api_key_path, provider.api_key.as_ref());
-
-    let oauth_path = format!("{field_prefix}.oauth_access_token");
-    push_secret_ref_observation(
-        observations,
-        oauth_path,
-        provider.oauth_access_token.as_ref(),
-    );
-
-    collect_provider_header_secret_observations(field_prefix, provider, observations);
-}
-
-fn collect_provider_header_secret_observations(
-    field_prefix: &str,
-    provider: &mvp::config::ProviderConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    for (header_name, header_value) in &provider.headers {
-        if !provider_header_may_contain_secret(header_name.as_str()) {
-            continue;
-        }
-
-        let field_path = format!("{field_prefix}.headers.{header_name}");
-        push_string_secret_observation(observations, field_path, Some(header_value.as_str()));
-    }
-}
-
-fn provider_header_may_contain_secret(header_name: &str) -> bool {
-    PROVIDER_SECRET_HEADER_NAMES
-        .iter()
-        .any(|candidate| header_name.eq_ignore_ascii_case(candidate))
-}
-
-fn collect_web_search_secret_observations(
-    config: &mvp::config::LoongConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let brave_path = "tools.web_search.brave_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        brave_path,
-        config.tools.web_search.brave_api_key.as_deref(),
-    );
-
-    let tavily_path = "tools.web_search.tavily_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        tavily_path,
-        config.tools.web_search.tavily_api_key.as_deref(),
-    );
-
-    let perplexity_path = "tools.web_search.perplexity_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        perplexity_path,
-        config.tools.web_search.perplexity_api_key.as_deref(),
-    );
-
-    let exa_path = "tools.web_search.exa_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        exa_path,
-        config.tools.web_search.exa_api_key.as_deref(),
-    );
-
-    let firecrawl_path = "tools.web_search.firecrawl_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        firecrawl_path,
-        config.tools.web_search.firecrawl_api_key.as_deref(),
-    );
-
-    let jina_path = "tools.web_search.jina_api_key".to_owned();
-    push_string_secret_observation(
-        observations,
-        jina_path,
-        config.tools.web_search.jina_api_key.as_deref(),
-    );
-}
-
-fn collect_channel_secret_observations(
-    config: &mvp::config::LoongConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    collect_telegram_secret_observations(&config.telegram, observations);
-    collect_feishu_secret_observations(&config.feishu, observations);
-    collect_matrix_secret_observations(&config.matrix, observations);
-    collect_wecom_secret_observations(&config.wecom, observations);
-    collect_discord_secret_observations(&config.discord, observations);
-    collect_line_secret_observations(&config.line, observations);
-    collect_dingtalk_secret_observations(&config.dingtalk, observations);
-    collect_webhook_secret_observations(&config.webhook, observations);
-    collect_email_secret_observations(&config.email, observations);
-    collect_slack_secret_observations(&config.slack, observations);
-    collect_google_chat_secret_observations(&config.google_chat, observations);
-    collect_mattermost_secret_observations(&config.mattermost, observations);
-    collect_nextcloud_talk_secret_observations(&config.nextcloud_talk, observations);
-    collect_synology_chat_secret_observations(&config.synology_chat, observations);
-    collect_teams_secret_observations(&config.teams, observations);
-    collect_imessage_secret_observations(&config.imessage, observations);
-    collect_whatsapp_secret_observations(&config.whatsapp, observations);
-}
-
-fn collect_telegram_secret_observations(
-    config: &mvp::config::TelegramChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bot_token_path = "telegram.bot_token".to_owned();
-    push_secret_ref_observation(observations, bot_token_path, config.bot_token.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let bot_token_path = format!("telegram.accounts.{account_id}.bot_token");
-        push_secret_ref_observation(observations, bot_token_path, account.bot_token.as_ref());
-    }
-}
-
-fn collect_feishu_secret_observations(
-    config: &mvp::config::FeishuChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let app_id_path = "feishu.app_id".to_owned();
-    push_secret_ref_observation(observations, app_id_path, config.app_id.as_ref());
-    let app_secret_path = "feishu.app_secret".to_owned();
-    push_secret_ref_observation(observations, app_secret_path, config.app_secret.as_ref());
-    let verification_path = "feishu.verification_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        verification_path,
-        config.verification_token.as_ref(),
-    );
-    let encrypt_key_path = "feishu.encrypt_key".to_owned();
-    push_secret_ref_observation(observations, encrypt_key_path, config.encrypt_key.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let app_id_path = format!("feishu.accounts.{account_id}.app_id");
-        push_secret_ref_observation(observations, app_id_path, account.app_id.as_ref());
-        let app_secret_path = format!("feishu.accounts.{account_id}.app_secret");
-        push_secret_ref_observation(observations, app_secret_path, account.app_secret.as_ref());
-        let verification_path = format!("feishu.accounts.{account_id}.verification_token");
-        push_secret_ref_observation(
-            observations,
-            verification_path,
-            account.verification_token.as_ref(),
-        );
-        let encrypt_key_path = format!("feishu.accounts.{account_id}.encrypt_key");
-        push_secret_ref_observation(observations, encrypt_key_path, account.encrypt_key.as_ref());
-    }
-}
-
-fn collect_matrix_secret_observations(
-    config: &mvp::config::MatrixChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let access_token_path = "matrix.access_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        access_token_path,
-        config.access_token.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let access_token_path = format!("matrix.accounts.{account_id}.access_token");
-        push_secret_ref_observation(
-            observations,
-            access_token_path,
-            account.access_token.as_ref(),
-        );
-    }
-}
-
-fn collect_wecom_secret_observations(
-    config: &mvp::config::WecomChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bot_id_path = "wecom.bot_id".to_owned();
-    push_secret_ref_observation(observations, bot_id_path, config.bot_id.as_ref());
-    let secret_path = "wecom.secret".to_owned();
-    push_secret_ref_observation(observations, secret_path, config.secret.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let bot_id_path = format!("wecom.accounts.{account_id}.bot_id");
-        push_secret_ref_observation(observations, bot_id_path, account.bot_id.as_ref());
-        let secret_path = format!("wecom.accounts.{account_id}.secret");
-        push_secret_ref_observation(observations, secret_path, account.secret.as_ref());
-    }
-}
-
-fn collect_discord_secret_observations(
-    config: &mvp::config::DiscordChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bot_token_path = "discord.bot_token".to_owned();
-    push_secret_ref_observation(observations, bot_token_path, config.bot_token.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let bot_token_path = format!("discord.accounts.{account_id}.bot_token");
-        push_secret_ref_observation(observations, bot_token_path, account.bot_token.as_ref());
-    }
-}
-
-fn collect_line_secret_observations(
-    config: &mvp::config::LineChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let access_token_path = "line.channel_access_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        access_token_path,
-        config.channel_access_token.as_ref(),
-    );
-    let secret_path = "line.channel_secret".to_owned();
-    push_secret_ref_observation(observations, secret_path, config.channel_secret.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let access_token_path = format!("line.accounts.{account_id}.channel_access_token");
-        push_secret_ref_observation(
-            observations,
-            access_token_path,
-            account.channel_access_token.as_ref(),
-        );
-        let secret_path = format!("line.accounts.{account_id}.channel_secret");
-        push_secret_ref_observation(observations, secret_path, account.channel_secret.as_ref());
-    }
-}
-
-fn collect_dingtalk_secret_observations(
-    config: &mvp::config::DingtalkChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let webhook_path = "dingtalk.webhook_url".to_owned();
-    push_secret_ref_observation(observations, webhook_path, config.webhook_url.as_ref());
-    let secret_path = "dingtalk.secret".to_owned();
-    push_secret_ref_observation(observations, secret_path, config.secret.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let webhook_path = format!("dingtalk.accounts.{account_id}.webhook_url");
-        push_secret_ref_observation(observations, webhook_path, account.webhook_url.as_ref());
-        let secret_path = format!("dingtalk.accounts.{account_id}.secret");
-        push_secret_ref_observation(observations, secret_path, account.secret.as_ref());
-    }
-}
-
-fn collect_webhook_secret_observations(
-    config: &mvp::config::WebhookChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let endpoint_path = "webhook.endpoint_url".to_owned();
-    push_secret_ref_observation(observations, endpoint_path, config.endpoint_url.as_ref());
-    let auth_token_path = "webhook.auth_token".to_owned();
-    push_secret_ref_observation(observations, auth_token_path, config.auth_token.as_ref());
-    let signing_secret_path = "webhook.signing_secret".to_owned();
-    push_secret_ref_observation(
-        observations,
-        signing_secret_path,
-        config.signing_secret.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let endpoint_path = format!("webhook.accounts.{account_id}.endpoint_url");
-        push_secret_ref_observation(observations, endpoint_path, account.endpoint_url.as_ref());
-        let auth_token_path = format!("webhook.accounts.{account_id}.auth_token");
-        push_secret_ref_observation(observations, auth_token_path, account.auth_token.as_ref());
-        let signing_secret_path = format!("webhook.accounts.{account_id}.signing_secret");
-        push_secret_ref_observation(
-            observations,
-            signing_secret_path,
-            account.signing_secret.as_ref(),
-        );
-    }
-}
-
-fn collect_email_secret_observations(
-    config: &mvp::config::EmailChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let smtp_username_path = "email.smtp_username".to_owned();
-    push_secret_ref_observation(
-        observations,
-        smtp_username_path,
-        config.smtp_username.as_ref(),
-    );
-    let smtp_password_path = "email.smtp_password".to_owned();
-    push_secret_ref_observation(
-        observations,
-        smtp_password_path,
-        config.smtp_password.as_ref(),
-    );
-    let imap_username_path = "email.imap_username".to_owned();
-    push_secret_ref_observation(
-        observations,
-        imap_username_path,
-        config.imap_username.as_ref(),
-    );
-    let imap_password_path = "email.imap_password".to_owned();
-    push_secret_ref_observation(
-        observations,
-        imap_password_path,
-        config.imap_password.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let smtp_username_path = format!("email.accounts.{account_id}.smtp_username");
-        push_secret_ref_observation(
-            observations,
-            smtp_username_path,
-            account.smtp_username.as_ref(),
-        );
-        let smtp_password_path = format!("email.accounts.{account_id}.smtp_password");
-        push_secret_ref_observation(
-            observations,
-            smtp_password_path,
-            account.smtp_password.as_ref(),
-        );
-        let imap_username_path = format!("email.accounts.{account_id}.imap_username");
-        push_secret_ref_observation(
-            observations,
-            imap_username_path,
-            account.imap_username.as_ref(),
-        );
-        let imap_password_path = format!("email.accounts.{account_id}.imap_password");
-        push_secret_ref_observation(
-            observations,
-            imap_password_path,
-            account.imap_password.as_ref(),
-        );
-    }
-}
-
-fn collect_slack_secret_observations(
-    config: &mvp::config::SlackChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bot_token_path = "slack.bot_token".to_owned();
-    push_secret_ref_observation(observations, bot_token_path, config.bot_token.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let bot_token_path = format!("slack.accounts.{account_id}.bot_token");
-        push_secret_ref_observation(observations, bot_token_path, account.bot_token.as_ref());
-    }
-}
-
-fn collect_google_chat_secret_observations(
-    config: &mvp::config::GoogleChatChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let webhook_path = "google_chat.webhook_url".to_owned();
-    push_secret_ref_observation(observations, webhook_path, config.webhook_url.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let webhook_path = format!("google_chat.accounts.{account_id}.webhook_url");
-        push_secret_ref_observation(observations, webhook_path, account.webhook_url.as_ref());
-    }
-}
-
-fn collect_mattermost_secret_observations(
-    config: &mvp::config::MattermostChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bot_token_path = "mattermost.bot_token".to_owned();
-    push_secret_ref_observation(observations, bot_token_path, config.bot_token.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let bot_token_path = format!("mattermost.accounts.{account_id}.bot_token");
-        push_secret_ref_observation(observations, bot_token_path, account.bot_token.as_ref());
-    }
-}
-
-fn collect_nextcloud_talk_secret_observations(
-    config: &mvp::config::NextcloudTalkChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let shared_secret_path = "nextcloud_talk.shared_secret".to_owned();
-    push_secret_ref_observation(
-        observations,
-        shared_secret_path,
-        config.shared_secret.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let shared_secret_path = format!("nextcloud_talk.accounts.{account_id}.shared_secret");
-        push_secret_ref_observation(
-            observations,
-            shared_secret_path,
-            account.shared_secret.as_ref(),
-        );
-    }
-}
-
-fn collect_synology_chat_secret_observations(
-    config: &mvp::config::SynologyChatChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let token_path = "synology_chat.token".to_owned();
-    push_secret_ref_observation(observations, token_path, config.token.as_ref());
-    let incoming_url_path = "synology_chat.incoming_url".to_owned();
-    push_secret_ref_observation(
-        observations,
-        incoming_url_path,
-        config.incoming_url.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let token_path = format!("synology_chat.accounts.{account_id}.token");
-        push_secret_ref_observation(observations, token_path, account.token.as_ref());
-        let incoming_url_path = format!("synology_chat.accounts.{account_id}.incoming_url");
-        push_secret_ref_observation(
-            observations,
-            incoming_url_path,
-            account.incoming_url.as_ref(),
-        );
-    }
-}
-
-fn collect_teams_secret_observations(
-    config: &mvp::config::TeamsChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let webhook_path = "teams.webhook_url".to_owned();
-    push_secret_ref_observation(observations, webhook_path, config.webhook_url.as_ref());
-    let app_id_path = "teams.app_id".to_owned();
-    push_secret_ref_observation(observations, app_id_path, config.app_id.as_ref());
-    let app_password_path = "teams.app_password".to_owned();
-    push_secret_ref_observation(
-        observations,
-        app_password_path,
-        config.app_password.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let webhook_path = format!("teams.accounts.{account_id}.webhook_url");
-        push_secret_ref_observation(observations, webhook_path, account.webhook_url.as_ref());
-        let app_id_path = format!("teams.accounts.{account_id}.app_id");
-        push_secret_ref_observation(observations, app_id_path, account.app_id.as_ref());
-        let app_password_path = format!("teams.accounts.{account_id}.app_password");
-        push_secret_ref_observation(
-            observations,
-            app_password_path,
-            account.app_password.as_ref(),
-        );
-    }
-}
-
-fn collect_imessage_secret_observations(
-    config: &mvp::config::ImessageChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let bridge_token_path = "imessage.bridge_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        bridge_token_path,
-        config.bridge_token.as_ref(),
-    );
-
-    for (account_id, account) in &config.accounts {
-        let bridge_token_path = format!("imessage.accounts.{account_id}.bridge_token");
-        push_secret_ref_observation(
-            observations,
-            bridge_token_path,
-            account.bridge_token.as_ref(),
-        );
-    }
-}
-
-fn collect_whatsapp_secret_observations(
-    config: &mvp::config::WhatsappChannelConfig,
-    observations: &mut Vec<SecretObservation>,
-) {
-    let access_token_path = "whatsapp.access_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        access_token_path,
-        config.access_token.as_ref(),
-    );
-    let verify_token_path = "whatsapp.verify_token".to_owned();
-    push_secret_ref_observation(
-        observations,
-        verify_token_path,
-        config.verify_token.as_ref(),
-    );
-    let app_secret_path = "whatsapp.app_secret".to_owned();
-    push_secret_ref_observation(observations, app_secret_path, config.app_secret.as_ref());
-
-    for (account_id, account) in &config.accounts {
-        let access_token_path = format!("whatsapp.accounts.{account_id}.access_token");
-        push_secret_ref_observation(
-            observations,
-            access_token_path,
-            account.access_token.as_ref(),
-        );
-        let verify_token_path = format!("whatsapp.accounts.{account_id}.verify_token");
-        push_secret_ref_observation(
-            observations,
-            verify_token_path,
-            account.verify_token.as_ref(),
-        );
-        let app_secret_path = format!("whatsapp.accounts.{account_id}.app_secret");
-        push_secret_ref_observation(observations, app_secret_path, account.app_secret.as_ref());
-    }
-}
-
-fn push_string_secret_observation(
-    observations: &mut Vec<SecretObservation>,
-    field_path: String,
-    raw_value: Option<&str>,
-) {
-    let Some(raw_value) = raw_value else {
-        return;
-    };
-
-    let trimmed_value = raw_value.trim();
-    if trimmed_value.is_empty() {
-        return;
-    }
-
-    let secret_ref = SecretRef::Inline(trimmed_value.to_owned());
-    let kind = classify_secret_ref_kind(&secret_ref);
-    let Some(kind) = kind else {
-        return;
-    };
-
-    let observation = SecretObservation { field_path, kind };
-    observations.push(observation);
-}
-
-fn push_secret_ref_observation(
-    observations: &mut Vec<SecretObservation>,
-    field_path: String,
-    secret_ref: Option<&SecretRef>,
-) {
-    let Some(secret_ref) = secret_ref else {
-        return;
-    };
-
-    let kind = classify_secret_ref_kind(secret_ref);
-    let Some(kind) = kind else {
-        return;
-    };
-
-    let observation = SecretObservation { field_path, kind };
-    observations.push(observation);
-}
-
-fn classify_secret_ref_kind(secret_ref: &SecretRef) -> Option<SecretReferenceKind> {
-    match secret_ref {
-        SecretRef::Env { .. } => Some(SecretReferenceKind::Env),
-        SecretRef::File { .. } => Some(SecretReferenceKind::File),
-        SecretRef::Exec { .. } => Some(SecretReferenceKind::Exec),
-        SecretRef::Inline(_) => {
-            if secret_ref.inline_literal_value().is_some() {
-                return Some(SecretReferenceKind::InlineLiteral);
-            }
-
-            if secret_ref.explicit_env_name().is_some() {
-                return Some(SecretReferenceKind::Env);
-            }
-
-            None
-        }
-    }
-}
-
-fn summarize_secret_observations(observations: &[SecretObservation]) -> SecretObservationCounts {
-    let mut counts = SecretObservationCounts::default();
-
-    for observation in observations {
-        counts.record(observation.kind);
-    }
-
-    counts
-}
-
-fn observation_paths_for_kind(
-    observations: &[SecretObservation],
-    kind: SecretReferenceKind,
-) -> Vec<String> {
-    observations
-        .iter()
-        .filter(|observation| observation.kind == kind)
-        .map(|observation| observation.field_path.clone())
-        .collect()
 }
 
 fn render_shell_default_mode(
